refactor: EditedIssueNetworkTarget을 IssueNetworkTarget에 합치기 (bugpigg/issue-tracker#48)
- 불필요하게 NetworkTarget을 나누지 않고 하나의 NetworkTarget 열거형에서 처리할 수 있도록, IssueNetworkTarget의 case 중 하나로 EditedIssueNetworkTarget에 있던 케이스를 추가함
